refactor: shared_preferences 관련 패키지 리팩토링
