+++ packages/apis/shared_preferences_api/.metadata
@@ -1,0 +1,10 @@
+# This file tracks properties of this Flutter project.
+# Used by Flutter tool to assess capabilities and perform upgrades etc.
+#
+# This file should be version controlled and should not be manually edited.
+
+version:
+  revision: "ba393198430278b6595976de84fe170f553cc728"
+  channel: "stable"
+
+project_type: package
