+++ packages/apis/shared_preferences_api/LICENSE
@@ -1,0 +1,1 @@
+TODO: Add your license here.
