+++ packages/apis/shared_preferences_api/lib/shared_preferences_api.dart
@@ -1,0 +1,3 @@
+library shared_preferences_api;
+
+export 'src/shared_preferences_api_client.dart';
