+++ packages/apis/shared_preferences_api/test/shared_preferences_api_test.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_test/flutter_test.dart';
+
+import 'package:shared_preferences_api/shared_preferences_api.dart';
+
+void main() {}
