+++ pubspec.yaml
@@ -33,6 +33,8 @@ dependencies:
   flutter_staggered_grid_view: ^0.7.0
   boj_api:
     path: packages/apis/boj_api
+  shared_preferences_api:
+    path: packages/apis/shared_preferences_api
   solved_api:
     path: packages/apis/solved_api
   contest_repository:
